Moved JSON -> POJO conversion to separate class.
